requirements.txt
numpy==1.17.4
scipy==1.3.0
tensorflow-gpu==2.0.0
opencv_python_headless==4.1.1.26
